LayoutTests/fast/forms/textarea-percentage-dimensions-expected.txt

PASS heightChanged("height:70%;") is 0
PASS heightChanged("margin-bottom: 20%;") is 0
PASS heightChanged("margin-top:20%;") is 0
PASS heightChanged("height:2000px; max-height:80%;") is 0
PASS heightChanged("height:100px; min-height:100%;") is 0
PASS heightChanged("padding-bottom: 20%;") is 0
PASS heightChanged("padding-top: 20%;") is 0
PASS successfullyParsed is true

TEST COMPLETE
